+++ src/Bootstrap.php
@@ -19,7 +19,6 @@ use yii\i18n\PhpMessageSource;
 class Bootstrap implements BootstrapInterface
 {
     /**
-     * @param Application $app
      * @param Application<User> $app
      */
     public function bootstrap($app): void